Removing markdown linting
Honestly, this is more of a chore than a value with differing opinions on rules, which I can reconsider in the future.

- lines longer than 80 characters
- fighting with markdown linter regarding overriding raw HTML usage
- number of lines
- indentation of code blocks

Perhaps I will meet a soul that can help me tune this to my particular tastes but not going to bother within the scope of this repository for now.